# Task.App/TaskApp.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace Task.App
{
    
    // Defining field properties for the TaskApp class
    public class TaskApp
    {
        public Guid ID { get; set; }
        public string name { get; set; }
        public string description { get; set; }
       

    }
}