# WebApiProject/bin/Debug/netcoreapp3.1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Project/ProjID": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "View all tasks in the project",
        "operationId": "Project_GetAllTask",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAppStored"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Project/id": {
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Delete project from Database and its tasks",
        "operationId": "Project_DeleteProj",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProjectAppStored": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProjectApp"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "startdate": {
                "type": "string",
                "format": "date-time"
              },
              "completiondate": {
                "type": "string",
                "format": "date-time"
              },
              "status": {
                "type": "string",
                "nullable": true
              },
              "priority": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "ProjectApp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "task": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TaskAppStored"
            }
          }
        }
      },
      "TaskAppStored": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TaskApp"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "status": {
                "type": "string",
                "nullable": true
              },
              "priority": {
                "type": "integer",
                "format": "int32"
              },
              "projectAppStoredID": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "TaskApp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}